Removed helper class and restored logic into ChatController
