--- Explorer/Assets/DCL/Chat/ChatControllerConversationEventsHelper.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 02780854478781548b956d30b859ad9a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
